move native library functionality to native/
