+++ native/utils.js
