+++ ui/DesignSystem/Primitive/Input/Directory.svelte
@@ -1,5 +1,5 @@
 <script>
-  import { getDirectoryPath } from "../../../lib/nativeUtils.js";
+  import { getDirectoryPath } from "../../../../native/utils.js";
 
   import Button from "../Button.svelte";
   import TextInput from "./Text.svelte";
