+++ ui/lib/path.js
@@ -1,6 +1,6 @@
 import regexparam from "regexparam";
 import { DEFAULT_PROJECT_REVISION } from "../config.js";
-import { BLOB, TREE } from "./types.js";
+import { BLOB, TREE } from "../../native/types.js";
 
 export const search = _params => "/search";
 export const network = _params => "/network";
